feat: remove the eager tracker tree
The eager tree read like a safe fallback and was not one. Measured against the
default on three real services: identical output on two, and on the third it
documented 194 of 280 routes — 31% missing, with no warning — while running
1.6x slower. Across the fixture suite it resolved four wiring shapes
incorrectly. Its one remaining advantage was memory on small projects, which is
not worth carrying a second engine that silently drops a third of an API.

Deprecated in 0.5.8 (#420), removed here: internal/spec/tracker.go (1976
lines), the --legacy-tracker flag and its warning, EngineConfig.UseLazyTracker,
the UI's "Analysis engine" selector (header, config panel, spec view, store,
CSS), and the five cross-engine parity tests, whose oracle no longer exists —
every fixture involved keeps its own structural test, so the behaviour stays
covered; what is lost is the meter.

Four pieces of tracker.go outlived it and moved to argument_type.go, because
the lazy tree needs them: ArgumentType with its constants and classifier,
assignmentKey (the producer index is looked up by exactly that composition —
#275), and the getString pool helper. The eager NODE also had a second life as
a hand-assembled TrackerNodeInterface: production code that needs one for a
single edge now uses a three-line edgeNode, and the 21 test files that built
nodes directly keep the old struct as a TEST-ONLY stub in mock_tracker_test.go,
since the lazy node cannot serve that purpose — it exists only inside a
materialised tree and derives its key from the tree's interner.

Three things the removal surfaced, each measured rather than assumed:

  - the diagram's node enrichment (function names, argument classification,
    call paths) read the eager node's struct, so it never ran for the default
    engine: a pre-removal diagram contains ZERO functionName fields, and the
    same diagram after removal is byte-identical. Deleted as dead code, with
    #442 filed to restore it for the lazy tree, whose node already holds the
    facts;
  - diffing those diagrams turned up a pre-existing nondeterminism — two
    baseline runs swap a pair of node labels — filed as #443;
  - two tests drove the analysis from a checked-in metadata.yaml. That
    round-trip is lossy: enough for the eager tree, not for the lazy one, so
    converting them would have looked like a capability regression. With fresh
    metadata the same fixture resolves fully, so the handler-factory
    assertions moved to the generator level (TestTestdata_EchoHandlerFactory)
    and the insight test now drives the engine — both testing what ships.

Zero output drift, which is the whole requirement for a removal: all 108
fixtures generate byte-identically, and gitea is byte-identical at 894 paths.
Coverage RISES to 96.2% (floor 96); the floor bump follows in its own PR per
the ratchet's convention. Docs updated: README's flag table and engine section,
docs/PERFORMANCE.md (one engine, and the limits table no longer splits by
engine), TRACKER_REDESIGN.md gains a status header marking it as the record of a
completed migration, and CLAUDE.md's pipeline description.

Closes #425. Closes #410. Closes #286.
