+++ cmd/apispecui/assets/js/actions.js
@@ -250,7 +250,6 @@ function fullGenerateRequest() {
     exclude: c.exclude,
     overrides: c.overrides,
     frameworkConfig: s.frameworkConfig || undefined,
-    legacyTracker: !!s.legacyTracker,
     limits: s.limits || {},
   };
 }
