+++ cmd/apispecui/assets/js/store.js
@@ -11,7 +11,6 @@ const state = {
   framework: "net/http",
   supportedFrameworks: ["gin", "chi", "echo", "fiber", "mux", "net/http"],
   openapiVersion: "3.1.0",
-  legacyTracker: false, // analysis engine: false = lazy tracker (default), true = legacy eager
   // Tree-expansion limits. Empty means "use the engine default" — the server
   // fills any unset field, so the UI never has to hardcode a number that would
   // then drift from Go (issue #233).
@@ -62,7 +61,7 @@ const state = {
 
 // View prefs persisted across refreshes so a reload lands the user back where
 // they were (which tab, which spec viewer, panel layout).
-const PERSIST_KEYS = ["mode", "specView", "panelCollapsed", "legacyTracker", "limits"];
+const PERSIST_KEYS = ["mode", "specView", "panelCollapsed", "limits"];
 const PERSIST_LS_KEY = "apispecui.view";
 
 function loadPersistedView() {
