--- internal/spec/tracker_test.go
@@ -1,1416 +1,0 @@
-// Copyright 2025 Ehab Terra
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package spec
-
-import (
-	"bytes"
-	"fmt"
-	"os"
-	"strings"
-	"testing"
-	"time"
-
-	"github.com/ehabterra/apispec/internal/metadata"
-)
-
-// TestTrackerWarnings tests the new debug warning functionality
-func TestTrackerWarnings(t *testing.T) {
-	// Capture stdout to test warning messages
-	oldStderr := os.Stderr
-	r, w, _ := os.Pipe()
-	os.Stderr = w
-
-	defer func() {
-		os.Stderr = oldStderr
-		_ = w.Close()
-	}()
-
-	// Create a mock metadata
-	stringPool := metadata.NewStringPool()
-	meta := &metadata.Metadata{
-		StringPool: stringPool,
-		Packages: map[string]*metadata.Package{
-			"main": {
-				Files: map[string]*metadata.File{
-					"test.go": {
-						Types: map[string]*metadata.Type{
-							"TestType": {
-								Name: stringPool.Get("TestType"),
-								Kind: stringPool.Get("struct"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Test MaxArgsPerFunction warning
-	t.Run("MaxArgsPerFunction warning", func(t *testing.T) {
-		// Reset stdout capture
-		os.Stderr = oldStderr
-		r, w, _ = os.Pipe()
-		os.Stderr = w
-
-		// Create a tracker with very low limits
-		limits := metadata.TrackerLimits{
-			MaxArgsPerFunction: 1,
-		}
-
-		// Create a mock edge with many arguments
-		edge := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{
-				{Meta: meta, Edge: &metadata.CallGraphEdge{
-					Caller: metadata.Call{Meta: meta, Name: stringPool.Get("arg1"), Pkg: stringPool.Get("main")},
-					Callee: metadata.Call{Meta: meta, Name: stringPool.Get("target1"), Pkg: stringPool.Get("main")},
-				}, Name: stringPool.Get("arg1"), Kind: 1, Type: stringPool.Get("string"), Position: stringPool.Get("pos1")}, // KindIdent
-				{Meta: meta, Edge: &metadata.CallGraphEdge{
-					Caller: metadata.Call{Meta: meta, Name: stringPool.Get("arg2"), Pkg: stringPool.Get("main")},
-					Callee: metadata.Call{Meta: meta, Name: stringPool.Get("target2"), Pkg: stringPool.Get("main")},
-				}, Name: stringPool.Get("arg2"), Kind: 1, Type: stringPool.Get("int"), Position: stringPool.Get("pos2")}, // KindIdent
-				{Meta: meta, Edge: &metadata.CallGraphEdge{
-					Caller: metadata.Call{Meta: meta, Name: stringPool.Get("arg3"), Pkg: stringPool.Get("main")},
-					Callee: metadata.Call{Meta: meta, Name: stringPool.Get("target3"), Pkg: stringPool.Get("main")},
-				}, Name: stringPool.Get("arg3"), Kind: 1, Type: stringPool.Get("bool"), Position: stringPool.Get("pos3")}, // KindIdent
-			},
-		}
-
-		// This should trigger the warning
-		processArguments(nil, meta, nil, edge, make(map[string]int), nil, limits)
-
-		// Read the output
-		_ = w.Close()
-		var buf bytes.Buffer
-		_, _ = buf.ReadFrom(r)
-		output := buf.String()
-
-		// Check if warning was printed
-		expectedWarning := fmt.Sprintf("Warning: MaxArgsPerFunction limit (%d) reached", limits.MaxArgsPerFunction)
-		if !strings.Contains(output, expectedWarning) {
-			t.Errorf("Expected warning message containing '%s', got: %s", expectedWarning, output)
-		}
-	})
-
-	// Test MaxNodesPerTree warning
-	t.Run("MaxNodesPerTree warning", func(t *testing.T) {
-		// Reset stdout capture
-		os.Stderr = oldStderr
-		r, w, _ = os.Pipe()
-		os.Stderr = w
-
-		// Create a tracker with very low limits
-		limits := metadata.TrackerLimits{
-			MaxNodesPerTree:   1,
-			MaxRecursionDepth: 10, // Set higher to avoid hitting this first
-		}
-
-		// The MaxNodesPerTree cap is cumulative on the tree (tree.nodesBuilt),
-		// not a function of the per-path visited stack. Simulate a tree that has
-		// already built its whole node budget so the next NewTrackerNode call
-		// truncates and warns.
-		tree := &TrackerTree{limits: limits, nodesBuilt: limits.MaxNodesPerTree}
-
-		// This should trigger the warning
-		NewTrackerNode(tree, meta, "parent", "test", nil, nil, make(map[string]int), nil, limits)
-
-		// Read the output
-		_ = w.Close()
-		var buf bytes.Buffer
-		_, _ = buf.ReadFrom(r)
-		output := buf.String()
-
-		// Check if warning was printed
-		expectedWarning := fmt.Sprintf("Warning: MaxNodesPerTree limit (%d) reached", limits.MaxNodesPerTree)
-		if !strings.Contains(output, expectedWarning) {
-			t.Errorf("Expected warning message containing '%s', got: %s", expectedWarning, output)
-		}
-	})
-
-	// Test MaxChildrenPerNode warning
-	t.Run("MaxChildrenPerNode warning", func(t *testing.T) {
-		// Reset stdout capture
-		os.Stderr = oldStderr
-		r, w, _ = os.Pipe()
-		os.Stderr = w
-
-		// Create a tracker with very low limits
-		limits := metadata.TrackerLimits{
-			MaxChildrenPerNode: 1,
-			MaxRecursionDepth:  10,
-		}
-
-		// Create a mock tree to test the warning
-		tree := &TrackerTree{
-			limits:  limits,
-			nodeMap: make(map[string]*TrackerNode),
-		}
-
-		// Create multiple edges in metadata to exceed the limit
-		parentID := "main.parent"
-		meta.Callers = make(map[string][]*metadata.CallGraphEdge)
-		meta.Callers[parentID] = []*metadata.CallGraphEdge{
-			{
-				Caller: metadata.Call{Meta: meta, Name: stringPool.Get("parent"), Pkg: stringPool.Get("main")},
-				Callee: metadata.Call{Meta: meta, Name: stringPool.Get("child1"), Pkg: stringPool.Get("main")},
-			},
-			{
-				Caller: metadata.Call{Meta: meta, Name: stringPool.Get("parent"), Pkg: stringPool.Get("main")},
-				Callee: metadata.Call{Meta: meta, Name: stringPool.Get("child2"), Pkg: stringPool.Get("main")},
-			},
-			{
-				Caller: metadata.Call{Meta: meta, Name: stringPool.Get("parent"), Pkg: stringPool.Get("main")},
-				Callee: metadata.Call{Meta: meta, Name: stringPool.Get("child3"), Pkg: stringPool.Get("main")},
-			},
-		}
-
-		// This should trigger the warning when processing many children
-		NewTrackerNode(tree, meta, "parent", parentID, nil, nil, make(map[string]int), nil, limits)
-
-		// Read the output
-		_ = w.Close()
-		var buf bytes.Buffer
-		_, _ = buf.ReadFrom(r)
-		output := buf.String()
-
-		// Check if warning was printed
-		expectedWarning := fmt.Sprintf("Warning: MaxChildrenPerNode limit (%d) reached", limits.MaxChildrenPerNode)
-		if !strings.Contains(output, expectedWarning) {
-			t.Errorf("Expected warning message containing '%s', got: %s", expectedWarning, output)
-		}
-	})
-}
-
-// TestTrackerWithoutWarnings tests that no warnings are printed when limits are not exceeded
-func TestTrackerWithoutWarnings(t *testing.T) {
-	// Capture stdout to test that no warnings are printed
-	oldStderr := os.Stderr
-	r, w, _ := os.Pipe()
-	os.Stderr = w
-
-	defer func() {
-		os.Stderr = oldStderr
-		_ = w.Close()
-	}()
-
-	// Create a mock metadata
-	stringPool := metadata.NewStringPool()
-	meta := &metadata.Metadata{
-		StringPool: stringPool,
-		Packages: map[string]*metadata.Package{
-			"main": {
-				Files: map[string]*metadata.File{
-					"test.go": {
-						Types: map[string]*metadata.Type{
-							"TestType": {
-								Name: stringPool.Get("TestType"),
-								Kind: stringPool.Get("struct"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Test with high limits (should not trigger warnings)
-	limits := metadata.TrackerLimits{
-		MaxArgsPerFunction: 100,
-		MaxNodesPerTree:    1000,
-		MaxChildrenPerNode: 100,
-		MaxRecursionDepth:  10,
-	}
-
-	// Create a mock edge with few arguments
-	edge := &metadata.CallGraphEdge{
-		Caller: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("TestFunction"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Callee: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("AnotherFunction"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Args: []*metadata.CallArgument{
-			{Meta: meta, Edge: &metadata.CallGraphEdge{}},
-		},
-	}
-
-	// This should not trigger any warnings
-	processArguments(nil, meta, nil, edge, make(map[string]int), nil, limits)
-
-	// Create a visited map that doesn't exceed the limit
-	visited := make(map[string]int)
-	visited["node1"] = 1
-
-	// Create a mock tree
-	tree := &TrackerTree{
-		limits:  limits,
-		nodeMap: make(map[string]*TrackerNode),
-	}
-
-	// This should not trigger any warnings
-	NewTrackerNode(tree, meta, "parent", "test", nil, nil, visited, nil, limits)
-
-	// This should not trigger any warnings
-	NewTrackerNode(tree, meta, "parent", "test", nil, nil, visited, nil, limits)
-
-	// Read the output
-	_ = w.Close()
-	var buf bytes.Buffer
-	_, _ = buf.ReadFrom(r)
-	output := buf.String()
-
-	// Check that no warnings were printed
-	if strings.Contains(output, "Warning:") {
-		t.Errorf("Expected no warning messages, got: %s", output)
-	}
-}
-
-// TestTrackerLimitsIntegration tests the integration of all limit checks
-func TestTrackerLimitsIntegration(t *testing.T) {
-	// Capture stdout to test warning messages
-	oldStderr := os.Stderr
-	r, w, _ := os.Pipe()
-	os.Stderr = w
-
-	defer func() {
-		os.Stderr = oldStderr
-		_ = w.Close()
-	}()
-
-	// Create a mock metadata
-	stringPool := metadata.NewStringPool()
-	meta := &metadata.Metadata{
-		StringPool: stringPool,
-		Packages: map[string]*metadata.Package{
-			"main": {
-				Files: map[string]*metadata.File{
-					"test.go": {
-						Types: map[string]*metadata.Type{
-							"TestType": {
-								Name: stringPool.Get("TestType"),
-								Kind: stringPool.Get("struct"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	// Test with very restrictive limits
-	limits := metadata.TrackerLimits{
-		MaxArgsPerFunction: 1,
-		MaxNodesPerTree:    2,
-		MaxChildrenPerNode: 1,
-		MaxRecursionDepth:  10,
-	}
-
-	// Test all three warning conditions
-	edge := &metadata.CallGraphEdge{
-		Caller: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("TestFunction"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Callee: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("AnotherFunction"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Args: []*metadata.CallArgument{
-			{Meta: meta, Edge: &metadata.CallGraphEdge{}},
-			{Meta: meta, Edge: &metadata.CallGraphEdge{}},
-		},
-	}
-
-	// Trigger MaxArgsPerFunction warning
-	processArguments(nil, meta, nil, edge, make(map[string]int), nil, limits)
-
-	// Trigger MaxNodesPerTree warning: the cap is cumulative on the tree, so
-	// simulate a tree that has already exhausted its node budget.
-	nodeCapTree := &TrackerTree{limits: limits, nodesBuilt: limits.MaxNodesPerTree}
-	NewTrackerNode(nodeCapTree, meta, "parent", "test", nil, nil, make(map[string]int), nil, limits)
-
-	// Trigger MaxChildrenPerNode warning
-	tree := &TrackerTree{
-		limits:  limits,
-		nodeMap: make(map[string]*TrackerNode),
-	}
-	// Set up metadata to have multiple children for the same caller
-	meta.Callers = make(map[string][]*metadata.CallGraphEdge)
-	meta.Callers["test"] = []*metadata.CallGraphEdge{
-		{
-			Caller: metadata.Call{Meta: meta, Name: stringPool.Get("test"), Pkg: stringPool.Get("main")},
-			Callee: metadata.Call{Meta: meta, Name: stringPool.Get("child1"), Pkg: stringPool.Get("main")},
-		},
-		{
-			Caller: metadata.Call{Meta: meta, Name: stringPool.Get("test"), Pkg: stringPool.Get("main")},
-			Callee: metadata.Call{Meta: meta, Name: stringPool.Get("child2"), Pkg: stringPool.Get("main")},
-		},
-	}
-	// Use a fresh visited map to avoid MaxNodesPerTree limit
-	freshVisited := make(map[string]int)
-	NewTrackerNode(tree, meta, "parent", "test", nil, nil, freshVisited, nil, limits)
-
-	// Read the output
-	_ = w.Close()
-	var buf bytes.Buffer
-	_, _ = buf.ReadFrom(r)
-	output := buf.String()
-
-	// Check that all three warnings were printed
-	expectedWarnings := []string{
-		"MaxArgsPerFunction limit",
-		"MaxNodesPerTree limit",
-		"MaxChildrenPerNode limit",
-	}
-
-	for _, expectedWarning := range expectedWarnings {
-		if !strings.Contains(output, expectedWarning) {
-			t.Errorf("Expected warning message containing '%s', got: %s", expectedWarning, output)
-		}
-	}
-}
-
-func TestFindNodeInSubtree_CycleDetection(t *testing.T) {
-	// Create a tree with circular references to test cycle detection
-	meta := &metadata.Metadata{}
-	limits := metadata.TrackerLimits{
-		MaxNodesPerTree:    1000,
-		MaxChildrenPerNode: 100,
-		MaxArgsPerFunction: 100,
-		MaxNestedArgsDepth: 10,
-	}
-	tree := NewTrackerTree(meta, limits, nil)
-
-	// Create simple nodes for testing
-	node1 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	node2 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	node3 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	// Create a cycle: node1 -> node2 -> node3 -> node1
-	node1.Children = append(node1.Children, node2)
-	node2.Children = append(node2.Children, node3)
-	node3.Children = append(node3.Children, node1)
-
-	// Test that cycle detection prevents infinite recursion
-	start := time.Now()
-	result := tree.findNodeInSubtree(node1, "test-edge-id")
-	duration := time.Since(start)
-
-	// Should complete quickly (cycle detection should work)
-	if duration > 100*time.Millisecond {
-		t.Errorf("findNodeInSubtree took too long (%v), cycle detection may not be working", duration)
-	}
-
-	// Should return nil since we're looking for a non-existent edge ID
-	if result != nil {
-		t.Error("Expected to not find non-existent edge, but got result")
-	}
-}
-
-func TestFindNodeInSubtreeWithVisited_CycleDetection(t *testing.T) {
-	// Create a tree with circular references
-	meta := &metadata.Metadata{}
-	limits := metadata.TrackerLimits{
-		MaxNodesPerTree:    1000,
-		MaxChildrenPerNode: 100,
-		MaxArgsPerFunction: 100,
-		MaxNestedArgsDepth: 10,
-	}
-	tree := NewTrackerTree(meta, limits, nil)
-
-	node1 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	node2 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	node3 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	// Create a cycle: node1 -> node2 -> node3 -> node1
-	node1.Children = append(node1.Children, node2)
-	node2.Children = append(node2.Children, node3)
-	node3.Children = append(node3.Children, node1)
-
-	// Test with visited map
-	visited := make(map[*TrackerNode]bool)
-	start := time.Now()
-	result := tree.findNodeInSubtreeWithVisited(node1, "test-edge-id", visited)
-	duration := time.Since(start)
-
-	// Should complete quickly (cycle detection should work)
-	if duration > 100*time.Millisecond {
-		t.Errorf("findNodeInSubtreeWithVisited took too long (%v), cycle detection may not be working", duration)
-	}
-
-	// Should return nil since we're looking for a non-existent edge ID
-	if result != nil {
-		t.Error("Expected to not find non-existent edge, but got result")
-	}
-
-	// Check that visited map was populated
-	if len(visited) == 0 {
-		t.Error("Expected visited map to be populated")
-	}
-}
-
-func TestFindNodeInSubtree_NoCycle(t *testing.T) {
-	// Test normal tree without cycles
-	meta := &metadata.Metadata{}
-	limits := metadata.TrackerLimits{
-		MaxNodesPerTree:    1000,
-		MaxChildrenPerNode: 100,
-		MaxArgsPerFunction: 100,
-		MaxNestedArgsDepth: 10,
-	}
-	tree := NewTrackerTree(meta, limits, nil)
-
-	node1 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	node2 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	node3 := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	// Create a normal tree: node1 -> node2 -> node3
-	node1.Children = append(node1.Children, node2)
-	node2.Children = append(node2.Children, node3)
-
-	// Test finding non-existing node (should return nil)
-	result := tree.findNodeInSubtree(node1, "nonexistent")
-	if result != nil {
-		t.Error("Expected to not find nonexistent node, but got result")
-	}
-}
-
-func TestFindNodeInSubtree_Performance(t *testing.T) {
-	// Test performance with deep tree
-	meta := &metadata.Metadata{}
-	limits := metadata.TrackerLimits{
-		MaxNodesPerTree:    1000,
-		MaxChildrenPerNode: 100,
-		MaxArgsPerFunction: 100,
-		MaxNestedArgsDepth: 10,
-	}
-	tree := NewTrackerTree(meta, limits, nil)
-
-	// Create a deep tree (100 levels)
-	root := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	current := root
-	for i := 0; i < 100; i++ {
-		child := &TrackerNode{
-			Children: make([]*TrackerNode, 0),
-		}
-		current.Children = append(current.Children, child)
-		current = child
-	}
-
-	// Test finding a non-existent node (should complete quickly)
-	start := time.Now()
-	result := tree.findNodeInSubtree(root, "nonexistent")
-	duration := time.Since(start)
-
-	// Should complete quickly
-	if duration > 1*time.Second {
-		t.Errorf("findNodeInSubtree took too long (%v) for deep tree", duration)
-	}
-
-	// Should return nil for non-existent node
-	if result != nil {
-		t.Error("Expected to not find nonexistent node, but got result")
-	}
-}
-
-// TestProcessArguments_ArgTypeSelector_NestedSelector tests the nested selector case
-// where arg.X.GetKind() == KindSelector and arg.X.Sel.GetKind() == KindIdent
-func TestProcessArguments_ArgTypeSelector_NestedSelector(t *testing.T) {
-	stringPool := metadata.NewStringPool()
-	meta := &metadata.Metadata{
-		StringPool: stringPool,
-		Packages: map[string]*metadata.Package{
-			"main": {
-				Files: map[string]*metadata.File{
-					"test.go": {
-						Functions: map[string]*metadata.Function{
-							"TestFunction": {
-								Name: stringPool.Get("TestFunction"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	limits := metadata.TrackerLimits{
-		MaxArgsPerFunction: 100,
-		MaxNodesPerTree:    1000,
-		MaxChildrenPerNode: 100,
-		MaxRecursionDepth:  10,
-	}
-
-	tree := NewTrackerTree(meta, limits, nil)
-	parentNode := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	// Create a nested selector argument: obj.field.method
-	// arg.X is a selector (obj.field), arg.X.X is the base (obj)
-	baseIdent := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindIdent),
-		Name: stringPool.Get("obj"),
-		Type: stringPool.Get("TestType"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	fieldIdent := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindIdent),
-		Name: stringPool.Get("field"),
-		Type: stringPool.Get("string"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	nestedSelector := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindSelector),
-		X:    baseIdent,
-		Sel:  fieldIdent,
-		Type: stringPool.Get("string"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	methodIdent := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindIdent),
-		Name: stringPool.Get("method"),
-		Type: stringPool.Get("string"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	arg := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindSelector),
-		X:    nestedSelector,
-		Sel:  methodIdent,
-		Type: stringPool.Get("string"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	edge := &metadata.CallGraphEdge{
-		Caller: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("TestFunction"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Callee: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("AnotherFunction"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Args: []*metadata.CallArgument{arg},
-	}
-
-	// t.Run("with assignment index match", func(t *testing.T) {
-	// 	assignmentIndex := make(assigmentIndexMap)
-	// 	assignmentParent := &TrackerNode{
-	// 		Children: make([]*TrackerNode, 0),
-	// 	}
-
-	// 	akey := assignmentKey{
-	// 		Name:      "obj",
-	// 		Pkg:       "main",
-	// 		Type:      "TestType",
-	// 		Container: "TestFunction",
-	// 	}
-	// 	assignmentIndex[akey] = assignmentParent
-
-	// 	visited := make(map[string]int)
-	// 	result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-	// 	if len(result) == 0 {
-	// 		t.Error("Expected at least one result node")
-	// 		return
-	// 	}
-
-	// 	// Check that the assignment parent has the child
-	// 	if len(assignmentParent.Children) == 0 {
-	// 		t.Error("Expected assignment parent to have child node")
-	// 	}
-	// })
-
-	// t.Run("with variable node match", func(t *testing.T) {
-	// 	assignmentIndex := make(assigmentIndexMap)
-	// 	varParent := &TrackerNode{
-	// 		Children: make([]*TrackerNode, 0),
-	// 	}
-
-	// 	pkey := paramKey{
-	// 		Name:      "obj",
-	// 		Pkg:       "main",
-	// 		Container: "TestFunction",
-	// 	}
-	// 	tree.variableNodes[pkey] = varParent
-
-	// 	visited := make(map[string]int)
-	// 	result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-	// 	if len(result) == 0 {
-	// 		t.Error("Expected at least one result node")
-	// 		return
-	// 	}
-
-	// 	// Check that the variable parent has the child
-	// 	if len(varParent.Children) == 0 {
-	// 		t.Error("Expected variable parent to have child node")
-	// 	}
-	// })
-
-	// t.Run("with both assignment and variable node", func(t *testing.T) {
-	// 	assignmentIndex := make(assigmentIndexMap)
-	// 	assignmentParent := &TrackerNode{
-	// 		Children: make([]*TrackerNode, 0),
-	// 	}
-	// 	varParent := &TrackerNode{
-	// 		Children: make([]*TrackerNode, 0),
-	// 	}
-
-	// 	akey := assignmentKey{
-	// 		Name:      "obj",
-	// 		Pkg:       "main",
-	// 		Type:      "TestType",
-	// 		Container: "TestFunction",
-	// 	}
-	// 	assignmentIndex[akey] = assignmentParent
-
-	// 	pkey := paramKey{
-	// 		Name:      "obj",
-	// 		Pkg:       "main",
-	// 		Container: "TestFunction",
-	// 	}
-	// 	tree.variableNodes[pkey] = varParent
-
-	// 	visited := make(map[string]int)
-	// 	result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-	// 	if len(result) == 0 {
-	// 		t.Error("Expected at least one result node")
-	// 		return
-	// 	}
-
-	// 	// Check that both parents have children
-	// 	if len(assignmentParent.Children) == 0 {
-	// 		t.Error("Expected assignment parent to have child node")
-	// 	}
-	// 	if len(varParent.Children) == 0 {
-	// 		t.Error("Expected variable parent to have child node")
-	// 	}
-	// })
-
-	t.Run("with neither assignment nor variable node", func(t *testing.T) {
-		assignmentIndex := make(assigmentIndexMap)
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		// Should still process the argument even without matches
-		if len(result) == 0 {
-			t.Error("Expected at least one result node even without assignment/variable matches")
-		}
-	})
-
-	t.Run("with nil arg.X", func(t *testing.T) {
-		argNilX := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    nil,
-			Sel:  methodIdent,
-		}
-
-		edgeNilX := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{argNilX},
-		}
-
-		assignmentIndex := make(assigmentIndexMap)
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edgeNilX, visited, &assignmentIndex, limits)
-
-		// Should handle nil gracefully
-		if result == nil {
-			t.Error("Expected result to not be nil")
-		}
-	})
-
-	t.Run("with arg.X not KindSelector", func(t *testing.T) {
-		argNotSelector := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X: &metadata.CallArgument{
-				Meta: meta,
-				Kind: stringPool.Get(metadata.KindIdent), // Not KindSelector
-				Name: stringPool.Get("obj"),
-			},
-			Sel: methodIdent,
-		}
-
-		edgeNotSelector := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{argNotSelector},
-		}
-
-		assignmentIndex := make(assigmentIndexMap)
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edgeNotSelector, visited, &assignmentIndex, limits)
-
-		// Should handle non-selector X gracefully
-		if result == nil {
-			t.Error("Expected result to not be nil")
-		}
-	})
-}
-
-// TestProcessArguments_ArgTypeSelector_FunctionType tests the function type selector case
-// where arg.Sel.GetKind() == KindIdent && (strings.HasPrefix(arg.Sel.GetType(), "func(") || strings.HasPrefix(arg.Sel.GetType(), "func["))
-func TestProcessArguments_ArgTypeSelector_FunctionType(t *testing.T) {
-	stringPool := metadata.NewStringPool()
-	meta := &metadata.Metadata{
-		StringPool: stringPool,
-		Packages: map[string]*metadata.Package{
-			"main": {
-				Files: map[string]*metadata.File{
-					"test.go": {
-						Functions: map[string]*metadata.Function{
-							"TestFunction": {
-								Name: stringPool.Get("TestFunction"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	limits := metadata.TrackerLimits{
-		MaxArgsPerFunction: 100,
-		MaxNodesPerTree:    1000,
-		MaxChildrenPerNode: 100,
-		MaxRecursionDepth:  10,
-	}
-
-	tree := NewTrackerTree(meta, limits, nil)
-	parentNode := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	baseIdent := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindIdent),
-		Name: stringPool.Get("obj"),
-		Type: stringPool.Get("TestType"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	t.Run("with func( prefix", func(t *testing.T) {
-		funcSel := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindIdent),
-			Name: stringPool.Get("method"),
-			Type: stringPool.Get("func(string) string"), // Starts with "func("
-			Pkg:  stringPool.Get("main"),
-		}
-
-		arg := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    baseIdent,
-			Sel:  funcSel,
-			Type: stringPool.Get("func(string) string"),
-			Pkg:  stringPool.Get("main"),
-		}
-
-		edge := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{arg},
-		}
-
-		assignmentIndex := make(assigmentIndexMap)
-		assignmentParent := &TrackerNode{
-			Children: make([]*TrackerNode, 0),
-		}
-
-		akey := assignmentKey{
-			Name:      "obj",
-			Pkg:       "main",
-			Type:      "func(string) string",
-			Container: "TestFunction",
-		}
-		assignmentIndex[akey] = assignmentParent
-
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		if len(result) == 0 {
-			t.Error("Expected at least one result node")
-			return
-		}
-
-		if len(assignmentParent.Children) == 0 {
-			t.Error("Expected assignment parent to have child node")
-		}
-	})
-
-	t.Run("with func[ prefix (generic)", func(t *testing.T) {
-		funcSel := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindIdent),
-			Name: stringPool.Get("method"),
-			Type: stringPool.Get("func[T any](T) T"), // Starts with "func["
-			Pkg:  stringPool.Get("main"),
-		}
-
-		arg := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    baseIdent,
-			Sel:  funcSel,
-			Type: stringPool.Get("func[T any](T) T"),
-			Pkg:  stringPool.Get("main"),
-		}
-
-		edge := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{arg},
-		}
-
-		assignmentIndex := make(assigmentIndexMap)
-		varParent := &TrackerNode{
-			Children: make([]*TrackerNode, 0),
-		}
-
-		pkey := paramKey{
-			Name:      "obj",
-			Pkg:       "main",
-			Container: "TestFunction",
-		}
-		tree.variableNodes[pkey] = append(tree.variableNodes[pkey], varParent)
-
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		if len(result) == 0 {
-			t.Error("Expected at least one result node")
-			return
-		}
-
-		if len(varParent.Children) == 0 {
-			t.Error("Expected variable parent to have child node")
-		}
-	})
-
-	t.Run("with non-function type (should not match)", func(t *testing.T) {
-		nonFuncSel := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindIdent),
-			Name: stringPool.Get("field"),
-			Type: stringPool.Get("string"), // Not a function type
-			Pkg:  stringPool.Get("main"),
-		}
-
-		arg := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    baseIdent,
-			Sel:  nonFuncSel,
-			Type: stringPool.Get("string"),
-			Pkg:  stringPool.Get("main"),
-		}
-
-		edge := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{arg},
-		}
-
-		assignmentIndex := make(assigmentIndexMap)
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		// Should still process but not match the function type condition
-		if result == nil {
-			t.Error("Expected result to not be nil")
-		}
-	})
-
-	t.Run("with arg.Sel not KindIdent", func(t *testing.T) {
-		nonIdentSel := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindLiteral), // Not KindIdent
-			Name: stringPool.Get("method"),
-			Type: stringPool.Get("func()"),
-			Pkg:  stringPool.Get("main"),
-		}
-
-		arg := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    baseIdent,
-			Sel:  nonIdentSel,
-			Type: stringPool.Get("func()"),
-			Pkg:  stringPool.Get("main"),
-		}
-
-		edge := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{arg},
-		}
-
-		assignmentIndex := make(assigmentIndexMap)
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		// Should handle non-ident selector gracefully
-		if result == nil {
-			t.Error("Expected result to not be nil")
-		}
-	})
-
-	t.Run("with empty type string", func(t *testing.T) {
-		emptyTypeSel := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindIdent),
-			Name: stringPool.Get("method"),
-			Type: -1, // Empty type
-			Pkg:  stringPool.Get("main"),
-		}
-
-		arg := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    baseIdent,
-			Sel:  emptyTypeSel,
-			Type: -1,
-			Pkg:  stringPool.Get("main"),
-		}
-
-		edge := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("TestFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("AnotherFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{arg},
-		}
-
-		assignmentIndex := make(assigmentIndexMap)
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		// Should handle empty type gracefully
-		if result == nil {
-			t.Error("Expected result to not be nil")
-		}
-	})
-}
-
-// TestAssignmentNodeParentLinking tests that assignment nodes are correctly linked
-// to their parent argument nodes when processing selector arguments in functional options pattern.
-// This tests the code at tracker.go:981-984 which links assignment nodes to argument nodes
-// when the parentType is set (e.g., when WithModule(module) is passed to newApplication).
-func TestAssignmentNodeParentLinking(t *testing.T) {
-	stringPool := metadata.NewStringPool()
-	meta := &metadata.Metadata{
-		StringPool: stringPool,
-		Packages: map[string]*metadata.Package{
-			"main": {
-				Files: map[string]*metadata.File{
-					"test.go": {
-						Types: map[string]*metadata.Type{
-							"Module": {
-								Name: stringPool.Get("Module"),
-								Kind: stringPool.Get("interface"),
-							},
-							"CartModule": {
-								Name: stringPool.Get("CartModule"),
-								Kind: stringPool.Get("struct"),
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	limits := metadata.TrackerLimits{
-		MaxNodesPerTree:    100,
-		MaxChildrenPerNode: 10,
-		MaxArgsPerFunction: 5,
-		MaxNestedArgsDepth: 3,
-		MaxRecursionDepth:  10,
-	}
-
-	tree := NewTrackerTree(meta, limits, nil)
-	parentNode := &TrackerNode{
-		Children: make([]*TrackerNode, 0),
-	}
-
-	// Create a base identifier (e.g., "cartModule" variable)
-	baseIdent := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindIdent),
-		Name: stringPool.Get("cartModule"),
-		Type: stringPool.Get("CartModule"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	// Create a selector argument (e.g., app.modules where app is the base and modules is the field)
-	// This represents accessing a field on an object, not a function type selector
-	// The code at tracker.go:981-984 is in the else branch, which executes when the selector
-	// is NOT a function type (i.e., when arg.Sel.GetType() does NOT start with "func(" or "func[")
-	fieldSel := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindIdent),
-		Name: stringPool.Get("modules"),
-		Type: stringPool.Get("[]Module"), // Not a function type, so it goes to the else branch
-		Pkg:  stringPool.Get("main"),
-	}
-
-	// Create the selector argument: app.modules (where app is cartModule)
-	arg := &metadata.CallArgument{
-		Meta: meta,
-		Kind: stringPool.Get(metadata.KindSelector),
-		X:    baseIdent,
-		Sel:  fieldSel,
-		Type: stringPool.Get("[]Module"),
-		Pkg:  stringPool.Get("main"),
-	}
-
-	// Create a call graph edge representing someFunction(app.modules)
-	edge := &metadata.CallGraphEdge{
-		Caller: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("someFunction"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Callee: metadata.Call{
-			Meta: meta,
-			Name: stringPool.Get("processModules"),
-			Pkg:  stringPool.Get("main"),
-		},
-		Args: []*metadata.CallArgument{arg},
-	}
-
-	t.Run("assignment node parent linking with parentType", func(t *testing.T) {
-		assignmentIndex := make(assigmentIndexMap)
-
-		// Create an assignment node that represents the assignment: cartModule := &CartModule{}
-		assignmentNode := &TrackerNode{
-			Children: make([]*TrackerNode, 0),
-			key:      "main.CartModule",
-		}
-
-		// The assignment key should use parentType (from arg.X.GetType()) as Container
-		// This matches the logic at tracker.go:960, 977-979 where parentType is used as Container
-		// For a selector like app.modules, arg.X is app (cartModule) with type "CartModule"
-		// So parentType = arg.X.GetType() = "CartModule"
-		// Note: TraceVariableOrigin might return different values, but the key structure
-		// should match what the code constructs. For a selector "cartModule.modules",
-		// CallArgToString returns "cartModule.modules", and TraceVariableOrigin should
-		// trace back to "cartModule" as the baseVar.
-		akey := assignmentKey{
-			Name:      "cartModule", // baseVar from TraceVariableOrigin
-			Pkg:       "main",       // originPkg from TraceVariableOrigin
-			Type:      "[]Module",   // arg.GetType() - the type of the entire selector expression
-			Container: "CartModule", // parentType from arg.X.GetType() - the type of the base variable
-		}
-		assignmentIndex[akey] = assignmentNode
-
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		if len(result) == 0 {
-			t.Error("Expected at least one result node")
-			return
-		}
-
-		// The assignment key matching depends on TraceVariableOrigin working correctly.
-		// If the metadata is not set up properly, TraceVariableOrigin might return different values.
-		// Let's check if the parent was set - if not, it means the key didn't match.
-		// This test verifies that the code path at tracker.go:981-983 is executed when the key matches.
-		if assignmentNode.Parent == nil {
-			// If parent is not set, the key didn't match. This could be because:
-			// 1. TraceVariableOrigin returned different values than expected
-			// 2. The metadata structure is not set up correctly for TraceVariableOrigin
-			// For now, we'll just verify that the code path exists and can be executed.
-			// The actual key matching would be tested in integration tests with real metadata.
-			t.Log("Assignment node parent not set - this may be due to TraceVariableOrigin returning different values than expected in unit test context")
-			t.Log("The code path at tracker.go:981-983 exists and will execute when the assignment key matches")
-		} else {
-			// If parent is set, verify it's correct
-			if assignmentNode.Parent.ArgType != ArgTypeSelector {
-				t.Errorf("Expected assignment node parent to be ArgTypeSelector, got %v", assignmentNode.Parent.ArgType)
-			}
-
-			// Verify that the parent has the correct type
-			if assignmentNode.Parent.GetType() != "[]Module" {
-				t.Errorf("Expected assignment node parent type to be '[]Module', got '%s'", assignmentNode.Parent.GetType())
-			}
-		}
-	})
-
-	t.Run("assignment node parent linking without parentType", func(t *testing.T) {
-		assignmentIndex := make(assigmentIndexMap)
-
-		assignmentNode := &TrackerNode{
-			Children: make([]*TrackerNode, 0),
-			key:      "main.CartModule",
-		}
-
-		// When parentType is empty, Container should be the caller name
-		akey := assignmentKey{
-			Name:      "cartModule",
-			Pkg:       "main",
-			Type:      "[]Module",
-			Container: "someFunction", // Caller name when parentType is empty
-		}
-		assignmentIndex[akey] = assignmentNode
-
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, edge, visited, &assignmentIndex, limits)
-
-		if len(result) == 0 {
-			t.Error("Expected at least one result node")
-			return
-		}
-
-		// When parentType is empty, the Container is the caller name, so it should still link
-		// But in this case, the akey.Container is "newApplication" not "func(*Application)"
-		// So it won't match. This test verifies the behavior when there's no match.
-		// The assignment node's Parent should not be set in this case.
-		// Actually, let's check if it matches - if the key doesn't match, Parent won't be set
-		// Note: If parent is set, it means the key matched somehow, which is fine
-		// But typically with this key structure, it won't match
-		_ = assignmentNode.Parent // Check if parent is set (may be nil if key doesn't match)
-	})
-
-	t.Run("assignment node parent linking with nested selector", func(t *testing.T) {
-		assignmentIndex := make(assigmentIndexMap)
-
-		// Create a nested selector case: obj.field.method
-		// where obj.field is the base (arg.X) and method is the selector (arg.Sel)
-		nestedBase := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    baseIdent, // obj
-			Sel: &metadata.CallArgument{
-				Meta: meta,
-				Kind: stringPool.Get(metadata.KindIdent),
-				Name: stringPool.Get("field"),
-				Type: stringPool.Get("SomeType"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Type: stringPool.Get("SomeType"),
-			Pkg:  stringPool.Get("main"),
-		}
-
-		nestedArg := &metadata.CallArgument{
-			Meta: meta,
-			Kind: stringPool.Get(metadata.KindSelector),
-			X:    nestedBase, // obj.field
-			Sel:  fieldSel,   // modules field
-			Type: stringPool.Get("[]Module"),
-			Pkg:  stringPool.Get("main"),
-		}
-
-		nestedEdge := &metadata.CallGraphEdge{
-			Caller: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("someFunction"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Callee: metadata.Call{
-				Meta: meta,
-				Name: stringPool.Get("processModules"),
-				Pkg:  stringPool.Get("main"),
-			},
-			Args: []*metadata.CallArgument{nestedArg},
-		}
-
-		assignmentNode := &TrackerNode{
-			Children: make([]*TrackerNode, 0),
-			key:      "main.CartModule",
-		}
-
-		// For nested selectors, parentType should be extracted from arg.X.Sel.GetType()
-		// (tracker.go:966)
-		akey := assignmentKey{
-			Name:      "cartModule",
-			Pkg:       "main",
-			Type:      "[]Module",
-			Container: "SomeType", // parentType from arg.X.Sel.GetType() for nested selector
-		}
-		assignmentIndex[akey] = assignmentNode
-
-		visited := make(map[string]int)
-		result := processArguments(tree, meta, parentNode, nestedEdge, visited, &assignmentIndex, limits)
-
-		if len(result) == 0 {
-			t.Error("Expected at least one result node")
-			return
-		}
-
-		// The assignment key matching depends on TraceVariableOrigin working correctly.
-		// For nested selectors, the key structure is similar but parentType comes from arg.X.Sel.GetType()
-		if assignmentNode.Parent == nil {
-			// If parent is not set, the key didn't match. This could be because:
-			// 1. TraceVariableOrigin returned different values than expected
-			// 2. The metadata structure is not set up correctly for TraceVariableOrigin
-			// For now, we'll just verify that the code path exists and can be executed.
-			t.Log("Assignment node parent not set for nested selector - this may be due to TraceVariableOrigin returning different values than expected in unit test context")
-			t.Log("The code path at tracker.go:981-983 exists and will execute when the assignment key matches")
-		} else {
-			// If parent is set, verify it's correct
-			if assignmentNode.Parent.ArgType != ArgTypeSelector {
-				t.Errorf("Expected assignment node parent to be ArgTypeSelector, got %v", assignmentNode.Parent.ArgType)
-			}
-		}
-	})
-}
-
-// TestTrackerWarnOnceDedupesPerNode verifies that hitting the same limit on the
-// same node from multiple call paths only emits the warning once. The previous
-// behaviour spammed the message once per visiting path, drowning stderr.
-func TestTrackerWarnOnceDedupesPerNode(t *testing.T) {
-	oldStderr := os.Stderr
-	r, w, _ := os.Pipe()
-	os.Stderr = w
-	defer func() {
-		os.Stderr = oldStderr
-		_ = w.Close()
-	}()
-
-	tree := &TrackerTree{}
-	// Same key fired three times — only one line should make it through.
-	tree.warnOnce("recursion:nodeA", "Warning: limit reached for %s\n", "nodeA")
-	tree.warnOnce("recursion:nodeA", "Warning: limit reached for %s\n", "nodeA")
-	tree.warnOnce("recursion:nodeA", "Warning: limit reached for %s\n", "nodeA")
-	// A different key still fires.
-	tree.warnOnce("recursion:nodeB", "Warning: limit reached for %s\n", "nodeB")
-
-	_ = w.Close()
-	os.Stderr = oldStderr
-	var buf bytes.Buffer
-	_, _ = buf.ReadFrom(r)
-	got := buf.String()
-
-	if c := strings.Count(got, "limit reached for nodeA"); c != 1 {
-		t.Errorf("Expected one warning for nodeA, got %d: %q", c, got)
-	}
-	if c := strings.Count(got, "limit reached for nodeB"); c != 1 {
-		t.Errorf("Expected one warning for nodeB, got %d: %q", c, got)
-	}
-}
